dashbord, course's chapters integration

+++ src/app/category/categories/categories.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 
 import { CategoryService } from '../category.service';
 import { Category } from '../category';
+declare var $: any;
+
+
 @Component({
   selector: 'app-categories',
   templateUrl: './categories.component.html',
@@ -21,9 +24,12 @@ export class CategoriesComponent implements OnInit {
       .subscribe(categories => this.categories = categories);
   }
 
-  delete(category: Category): void{
-    this.categories = this.categories.filter(c => c !== category);
-    this.categoryService.deleteCategory(category).subscribe();
+  delete(category: Category): void {
+    if (confirm('are you sure ?')) {
+
+      this.categories = this.categories.filter(c => c !== category);
+      this.categoryService.deleteCategory(category).subscribe();
+    }
   }
 
 }
